+++ tests/org.preesm.tests.integration/resources/org.preesm.tests.testHardware/Code/include/Matrix_Show.h
@@ -1,0 +1,26 @@
+/*
+	============================================================================
+	Name        : Matrix_Show.h
+	Author      : Leonardo Suriano
+	Version     : 1.0
+	Copyright   : UPM
+	Description :
+	============================================================================
+*/
+
+#ifndef MATRIX_SHOW_H
+#define MATRIX_SHOW_H
+
+#include "preesm.h"
+#include <stdio.h>
+#include <stdlib.h>
+
+
+
+/**
+* just it prints the elemtns of the matrix
+*
+*/
+void Matrix_Show(int n_cols, int n_rows, OUT int *matrix);
+
+#endif //MATRIX_SHOW_H