+++ tests/org.preesm.tests.integration/resources/org.preesm.tests.testHardware/Code/include/Matrix_Verification.h
@@ -1,0 +1,15 @@
+#ifndef MATRIX_VERIFICATION_H
+#define MATRIX_VERIFICATION_H
+
+#include "preesm.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <time.h>     // time()
+
+#define MEAN_LOOP_SIZE (1)
+#define VERIFY_RESULTS
+//#define VERBOSE
+//int matrix_verification_of2(int n_rows, int n_cols, IN int* A, OUT int* B);
+void matrix_verification(int n_rows, int n_cols, IN int* matrix_hw);
+
+#endif //MATRIX_VERIFICATION_H