+++ tests/org.preesm.tests.integration/resources/org.preesm.tests.testHardware/Code/include/artico3_accelerator_setup.h
@@ -1,0 +1,16 @@
+#ifndef _ARTICO3_HW_SETUP_
+#define _ARTICO3_HW_SETUP_
+
+#define SIZE_MEM_HW 49152   		//MemBytes
+#define MAX_NACCS 3 				//Number of acceletators
+#define N_MEMORY_BANKS 3			//MemBanks
+#define N_REGISTERS 1				//Regs
+#define A3_ACCELERATOR_REG_0 0x00	//
+#define INPUT A3_P_I				//matching PREESM definition
+#define OUTPUT A3_P_O				//matching PREESM definition
+
+//temp macro for testing purpose
+#define lsize_TO_BE_CHANGED 8
+#define gsize_TO_BE_CHANGED 8
+
+#endif // _ARTICO3_HW_SETUP_